requirements.txt
requests
beautifulsoup4
openpyxl
python-docx